+++ opentracing-spring-web-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 io.opentracing.contrib.spring.web.autoconfig.TracerAutoConfiguration,\
+io.opentracing.contrib.spring.web.autoconfig.TracerRegisterAutoConfiguration,\
 io.opentracing.contrib.spring.web.autoconfig.ServerTracingAutoConfiguration,\
 io.opentracing.contrib.spring.web.autoconfig.RestTemplateAutoConfiguration,\
 io.opentracing.contrib.spring.web.autoconfig.AsyncRestTemplateAutoConfiguration
